requirements.txt
Django==4.2.7
python-dotenv==1.0.0
django-crispy-forms==2.0
crispy-bootstrap5==0.7
psycopg2-binary==2.9.10
vnstock==3.1.0.2
python-decouple==3.8
gunicorn==23.0.0
whitenoise==6.9.0
annotated-types==0.7.0
asgiref==3.8.1
attrs==25.1.0
autobahn==24.4.2
Automat==24.8.1
beautifulsoup4==4.13.3
Brotli==1.1.0
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
constantly==23.10.4
contourpy==1.3.1
cryptography==44.0.2
cycler==0.12.1
daphne==4.1.2
dj-database-url==2.3.0
djangorestframework==3.15.2
et_xmlfile==2.0.0
fake-useragent==2.0.3
fonttools==4.56.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.0
kafka-python==2.0.2
kiwisolver==1.4.8
msgpack==1.1.0
numpy==2.2.3
openpyxl==3.1.5
service-identity==24.2.0
sqlparse==0.5.3
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
zope.interface==7.2
authlib==1.2.1
requests==2.31.0
# vnai==2.0.4
# vnstock_ezchart==0.0.2
# django-environ==0.11.2
# matplotlib==3.10.1
# channels==4.2.0
# channels_redis==4.2.1
# setuptools==76.0.0
# seaborn==0.13.2
# squarify==0.4.4